+++ CHANGELOG.md
@@ -17,7 +17,7 @@
  
 1.4.0
  - validate predit-fn results
- - new ns `scicloj.metamorph.ml.column-metric with metric-fns operating on datasets, not seq-of-numbers
+ - new ns `scicloj.metamorph.ml.column-metric` with metric-fns operating on datasets, not on seq-of-numbers
  - TMD's of rdatasets get doc-url as dataset name, ready for clicking [#36](https://github.com/scicloj/metamorph.ml/issues/36)
  - added model names to fastmath/ols [#24](https://github.com/scicloj/metamorph.ml/issues/24)
  - add support for fastmath GLM model [46](https://github.com/scicloj/metamorph.ml/issues/46)